+++ x/xrapid/iter.go
@@ -12,14 +12,38 @@ import (
 func SampledFromSeq[T any](seq iter.Seq[T]) *rapid.Generator[T] {
 	return rapid.Custom(
 		func(t *rapid.T) T {
-			slice := slices.Collect(seq)
+			values := slices.Collect(seq)
 
-			if len(slice) == 0 {
+			if len(values) == 0 {
 				t.Skip("sequence is empty")
 			}
 
-			index := rapid.IntRange(0, len(slice)-1).Draw(t, "index")
-			return slice[index]
+			return rapid.SampledFrom(values).Draw(t, "value")
+		},
+	)
+}
+
+// Pair is a key/value pair.
+type Pair[K, V any] struct {
+	Key   K
+	Value V
+}
+
+// SampledFromSeq2 returns a generator that produces random elements from the
+// given sequence.
+func SampledFromSeq2[K, V any](seq iter.Seq2[K, V]) *rapid.Generator[Pair[K, V]] {
+	return rapid.Custom(
+		func(t *rapid.T) Pair[K, V] {
+			var pairs []Pair[K, V]
+			for k, v := range seq {
+				pairs = append(pairs, Pair[K, V]{k, v})
+			}
+
+			if len(pairs) == 0 {
+				t.Skip("sequence is empty")
+			}
+
+			return rapid.SampledFrom(pairs).Draw(t, "pair")
 		},
 	)
 }